Renaming deployment module to interpreters.
